First Push
Site takes in an image, stores it in the sql database, moves a copy to
the websites images, and returns the name of where it is currently
stored.
With this it takes the url gives it to clarifi’s api and returns all
the possible ingredients found in the image.
Next is to send that data to the Spoonacular Api to get related meals.