API Call To Spoonacular
Clarifi API gives back all possible image results, then a new string is
created with all ingredients with a 85% chance of being in the original
dish. This string is used as part of the query string passed to the the
Spoonacular API to which it grabs the top results of meals that are
based off the same or similar ingredients.