Updates
Styling and color palette has slightly changed to a lighter feel.
List now populates when the user uploads an image and try to return at
most 10 possible meals that relate.

Need to create the modal for each one for last part:
As of now all the dynamically created meals are tagged 0-n for their id
so that one can click on it and then be presented with that
corresponding JSON array data that was passed back from the Spoonacular
result.