Update:
Site now creates modal and replaces content of said modal with the
selected JSON data based on which item in the similar dish list was
clicked.
Next is to call API one more time to get the cooking instructions and
the time, serving size of each meal which will be called once based of
the ids from the original API call that got all the similar dishes.
Then restyle the file upload button and done.