Update:
Restructured some javascript code to compartmentalize code into more
(smaller / useful) functions.
Second call to the Spoonacular API given the top meals id’s so that it
returns more data about those specific meals as a JSON file.

Will use this to add the instructions on how to cook the meal to the
modal.